+++ schema/src/test/scala/net/reactivecore/cjs/validator/ValidationSuite2019_09Spec.scala
@@ -1,3 +1,5 @@
 package net.reactivecore.cjs.validator
 
-class ValidationSuite2019_09Spec extends ValidationSuiteBase("draft2019-09")
+import net.reactivecore.cjs.Vocabularies
+
+class ValidationSuite2019_09Spec extends ValidationSuiteBase("draft2019-09", Vocabularies.vocabulary2019)
