+++ schema/src/test/scala/net/reactivecore/cjs/validator/ValidationSuite2020_12Spec.scala
@@ -1,3 +1,5 @@
 package net.reactivecore.cjs.validator
 
-class ValidationSuite2020_12Spec extends ValidationSuiteBase("draft2020-12")
+import net.reactivecore.cjs.Vocabularies
+
+class ValidationSuite2020_12Spec extends ValidationSuiteBase("draft2020-12", Vocabularies.vocabulary2020)
